Add collision detection for balls and paddles
Adds collision detection using the avian2d libary's ColliderAabb, and
correct the trajectory of the ball when a collision is made. This is
done by changing the values in the velocity component of the ball.

Now that I think about it the velocities direction should be a normal
vector in order for the translation to be consistent in all directions.
This is something that will need to be added later.