Add arena walls and check collisions.
Adds the arens walls such that the ball is blocked into a playable area.
When the ball hits one of the arena walls that the player paddles are
protecting the opposing players score will increase. At the moment if
a player reaches 5 goals, an info message is output the stdout.

An issue was found with the collisions with the paddles that is not
fixed in this patch. If the paddle is moving towards the ball when the
ball collides, the collision detection causes a panic because the balls
centre is now too close to the paddle to handle collision detection
properly. I'll fix this later. I'm more interested in getting features
working at the moment.